Loaded into CSV and deleted some columns

+++ following.py
@@ -5,6 +5,8 @@ import json
 import time
 import sys
 import numpy as np
+import pandas as pd
+import csv
 
 def save_cred(filename="my_creds.json"):
     """
@@ -71,28 +73,32 @@ def get_following(api):
     :param api:
     :return:
     """
+    filename = os.path.join(os.getcwd(), "following.json")
+    if not os.path.exists(filename):
+        # getting the friends list
+        friends = api.get_friend_ids()
+        print(" This account follows: {} accounts".format((len(friends))))
 
-    # getting the friends list
-    friends = api.get_friend_ids()
-    print(" This account follows: {} accounts".format((len(friends))))
+        batch_len = 100
+        num_batches = np.ceil(len(friends) / 100)
+        batches = (friends[i:i + batch_len] for i in range(0, len(friends), batch_len))
+        all_data = []
+        for batch_count, batch in enumerate(batches):
+            sys.stdout.write("\r")
+            sys.stdout.flush()
+            sys.stdout.write("Fetching batch: " + str(batch_count) + "/" + str(num_batches))
+            sys.stdout.flush()
+            users_list = api.lookup_users(user_id=batch)
+            users_json = (map(lambda t: t._json, users_list))
+            all_data += users_json
 
-    batch_len = 100
-    num_batches = np.ceil(len(friends) / 100)
-    batches = (friends[i:i + batch_len] for i in range(0, len(friends), batch_len))
-    all_data = []
-    for batch_count, batch in enumerate(batches):
-        sys.stdout.write("\r")
-        sys.stdout.flush()
-        sys.stdout.write("Fetching batch: " + str(batch_count) + "/" + str(num_batches))
-        sys.stdout.flush()
-        users_list = api.lookup_users(user_id=batch)
-        users_json = (map(lambda t: t._json, users_list))
-        all_data += users_json
+        # Save the credentials object to file
+        with open(filename, "w") as file:
+            json.dump(all_data, file)
+    else:
+        with open(filename, "r") as file:
+            all_data = json.load(file)
 
-
-    # Save the credentials object to file
-    with open((os.path.join(os.getcwd(), "following.json")), "w") as file:
-        json.dump(all_data, file)
     return all_data
 
 
@@ -104,3 +110,9 @@ if __name__ == "__main__":
     # the screen_name of the targeted user
     screen_name = "Nitin_wysiwyg"
     following = get_following(api)
+    print("Got the list")
+    keys = set().union(*(d.keys() for d in following))
+    with open(os.path.join(os.getcwd(),'following.csv'), 'w', encoding='utf-8', newline='')  as output_file:
+        dict_writer = csv.DictWriter(output_file, keys)
+        dict_writer.writeheader()
+        dict_writer.writerows(following)
